requirements.txt
openai>=1.0.0
anthropic>=0.20.0
python-dotenv
pydub
aiofiles
